move cron into cloudflare package
